refactor(termbud): separate subcommands and add tests
